refactor: add AttachmentConversionService interface and merging from the controller to the service
