+++ backend/src/main/java/eu/bbmri_eric/negotiator/negotiation/pdf/NegotiationPdfService.java
@@ -7,7 +7,8 @@ public interface NegotiationPdfService {
    *
    * @param negotiationId the id of the negotiation
    * @param templateName the name of the template to use for rendering
+   * @param includeAttachments whether to include attachments to generated PDF or not
    * @return a byte array representing the PDF document
    */
-  byte[] generatePdf(String negotiationId, String templateName);
+  byte[] generatePdf(String negotiationId, String templateName, boolean includeAttachments);
 }
